requirements.txt
google-generativeai
atlassian-python-api
langchain
langchain-google-genai
langgraph
chromadb
streamlit
python-dotenv
pydantic